feat: implement "What's New" feature with dynamic changelog display in webview footer
